Move helpers on separated file
