+++ src/mappings/factory.ts
@@ -1,7 +1,8 @@
-import { BigInt, BigDecimal, Address } from '@graphprotocol/graph-ts'
+import { BigInt, BigDecimal } from '@graphprotocol/graph-ts'
 import { LOG_NEW_POOL } from '../types/Factory/Factory'
 import { Balancer, Pool } from '../types/schema'
 import { Pool as PoolContract } from '../types/templates'
+import { ZERO_BD } from './helpers'
 
 export function handleNewPool(event: LOG_NEW_POOL): void {
   let factory = Balancer.load('1')
@@ -22,10 +23,10 @@ export function handleNewPool(event: LOG_NEW_POOL): void {
   pool.publicSwap = false
   pool.finalized = false
   pool.swapFee = BigDecimal.fromString('0.000001')
-  pool.totalWeight = BigDecimal.fromString('0')
-  pool.totalShares = BigDecimal.fromString('0')
-  pool.totalSwapVolume = BigDecimal.fromString('0')
-  pool.liquidity = BigDecimal.fromString('0')
+  pool.totalWeight = ZERO_BD
+  pool.totalShares = ZERO_BD
+  pool.totalSwapVolume = ZERO_BD
+  pool.liquidity = ZERO_BD
   pool.createTime = event.block.timestamp.toI32()
   pool.joinsCount = BigInt.fromI32(0)
   pool.exitsCount = BigInt.fromI32(0)
