+++ src/mappings/pool.ts
@@ -1,210 +1,24 @@
-import { BigInt, BigDecimal, Address, Bytes, ByteArray, log, store, dataSource } from '@graphprotocol/graph-ts'
+import { BigInt, Address, Bytes, store } from '@graphprotocol/graph-ts'
 import { LOG_CALL, LOG_JOIN, LOG_EXIT, LOG_SWAP, Transfer } from '../types/templates/Pool/Pool'
 import { BToken } from '../types/templates/Pool/BToken'
-import { BTokenBytes } from '../types/templates/Pool/BTokenBytes'
 import {
   Balancer,
   Pool,
   User,
   PoolToken,
   PoolShare,
-  Transaction,
   Swap,
   TokenPrice
 } from '../types/schema'
-
-let network = dataSource.network()
-let WETH: string;
-let USD: string;
-if (network == 'mainnet') {
-  WETH = '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2'
-  USD = '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48' // USDC
-} else if (network == 'kovan') {
-  WETH = '0xd0a1e359811322d97991e03f863a0c30c2cf029c'
-  USD = '0x1528f3fcc26d13f7079325fb78d9442607781c8c' // DAI
-}
-
-/************************************
- ********** Helpers ***********
- ************************************/
-
-function hexToDecimal(hexString: String, decimals: i32): BigDecimal {
-  let bytes = Bytes.fromHexString(hexString).reverse() as Bytes
-  let bi = BigInt.fromUnsignedBytes(bytes)
-  let scale = BigInt.fromI32(10).pow(decimals as u8).toBigDecimal()
-  return bi.divDecimal(scale)
-}
-
-function bigIntToDecimal(amount: BigInt, decimals: i32): BigDecimal {
-  let scale = BigInt.fromI32(10).pow(decimals as u8).toBigDecimal()
-  return amount.toBigDecimal().div(scale)
-}
-
-function tokenToDecimal(amount: BigDecimal, decimals: i32): BigDecimal {
-  let scale = BigInt.fromI32(10).pow(decimals as u8).toBigDecimal()
-  return amount.div(scale)
-}
-
-function createPoolShareEntity(id: string, pool: String, user: String): void {
-
-  let poolShare = new PoolShare(id)
-
-  let userdb = User.load(user)
-  if (userdb == null) {
-    userdb = new User(user)
-    userdb.save()
-  }
-  poolShare.userAddress = user
-  poolShare.poolId = pool
-  poolShare.balance = BigDecimal.fromString('0')
-  poolShare.save()
-}
-
-function createPoolTokenEntity(id: string, pool: String, address: String): void {
-  let token = BToken.bind(Address.fromString(address))
-  let tokenBytes = BTokenBytes.bind(Address.fromString(address))
-  let symbol = ''
-  let name = ''
-  let decimals = 18
-
-  // COMMENT THE LINES BELOW OUT FOR LOCAL DEV ON KOVAN
-
-  let symbolCall = token.try_symbol()
-  let nameCall = token.try_name()
-  let decimalCall = token.try_decimals()
-
-  if (symbolCall.reverted) {
-    let symbolBytesCall = tokenBytes.try_symbol()
-    if (!symbolBytesCall.reverted) {
-      symbol = symbolBytesCall.value.toString()
-    }
-  } else {
-    symbol = symbolCall.value
-  }
-
-  if (nameCall.reverted) {
-    let nameBytesCall = tokenBytes.try_name()
-    if (!nameBytesCall.reverted) {
-      name = nameBytesCall.value.toString()
-    }
-  } else {
-    name = nameCall.value
-  }
-
-  if (!decimalCall.reverted) {
-    decimals = decimalCall.value
-  }
-
-  // COMMENT THE LINES ABOVE OUT FOR LOCAL DEV ON KOVAN
-
-  // !!! COMMENT THE LINES BELOW OUT FOR NON-LOCAL DEPLOYMENT
-  // This code allows Symbols to be added when testing on local Kovan
-  /*
-  if(address == '0xd0a1e359811322d97991e03f863a0c30c2cf029c')
-    symbol = 'WETH';
-  else if(address == '0x1528f3fcc26d13f7079325fb78d9442607781c8c')
-    symbol = 'DAI'
-  else if(address == '0xef13c0c8abcaf5767160018d268f9697ae4f5375')
-    symbol = 'MKR'
-  else if(address == '0x2f375e94fc336cdec2dc0ccb5277fe59cbf1cae5')
-    symbol = 'USDC'
-  else if(address == '0x1f1f156e0317167c11aa412e3d1435ea29dc3cce')
-    symbol = 'BAT'
-  else if(address == '0x86436bce20258a6dcfe48c9512d4d49a30c4d8c4')
-    symbol = 'SNX'
-  else if(address == '0x8c9e6c40d3402480ace624730524facc5482798c')
-    symbol = 'REP'
-  */
-  // !!! COMMENT THE LINES ABOVE OUT FOR NON-LOCAL DEPLOYMENT
-
-  let poolToken = new PoolToken(id)
-  poolToken.poolId = pool
-  poolToken.address = address
-  poolToken.name = name
-  poolToken.symbol = symbol
-  poolToken.decimals = decimals
-  poolToken.balance = BigDecimal.fromString('0')
-  poolToken.denormWeight = BigDecimal.fromString('0')
-  poolToken.save()
-}
-
-function updatePoolLiquidity(id: string): void {
-  let pool = Pool.load(id)
-  let tokensList: Array<Bytes> = pool.tokensList
-
-  if (!tokensList || !pool.publicSwap) return
-
-  // Find pool liquidity
-
-  let hasPrice = false
-  let poolLiquidity = BigDecimal.fromString('0')
-
-  if (tokensList.includes(Address.fromString(USD))) {
-    let usdPoolTokenId = id.concat('-').concat(USD)
-    let usdPoolToken = PoolToken.load(usdPoolTokenId)
-    poolLiquidity = usdPoolToken.balance.div(usdPoolToken.denormWeight).times(pool.totalWeight)
-    hasPrice = true
-  } else if (tokensList.includes(Address.fromString(WETH))) {
-    let wethTokenPrice = TokenPrice.load(WETH)
-    if (wethTokenPrice !== null) {
-      let poolTokenId = id.concat('-').concat(WETH)
-      let poolToken = PoolToken.load(poolTokenId)
-      poolLiquidity = wethTokenPrice.price.times(poolToken.balance).div(poolToken.denormWeight).times(pool.totalWeight)
-      hasPrice = true
-    }
-  }
-
-  // Create or update token price
-
-  if (hasPrice) {
-    for (let i: i32 = 0; i < tokensList.length; i++) {
-      let tokenPriceId = tokensList[i].toHexString()
-      let tokenPrice = TokenPrice.load(tokenPriceId)
-      if (tokenPrice == null) {
-        tokenPrice = new TokenPrice(tokenPriceId)
-        tokenPrice.poolTokenId = ''
-        tokenPrice.poolLiquidity = BigDecimal.fromString('0')
-      }
-
-      let poolTokenId = id.concat('-').concat(tokenPriceId)
-      let poolToken = PoolToken.load(poolTokenId)
-
-      if (tokenPrice.poolTokenId == poolTokenId || poolLiquidity.gt(tokenPrice.poolLiquidity)) {
-        tokenPrice.price = BigDecimal.fromString('0')
-
-        if (poolToken.balance.gt(BigDecimal.fromString('0'))) {
-          tokenPrice.price = poolLiquidity.div(pool.totalWeight).times(poolToken.denormWeight).div(poolToken.balance)
-        }
-
-        tokenPrice.symbol = poolToken.symbol
-        tokenPrice.poolLiquidity = poolLiquidity
-        tokenPrice.poolTokenId = poolTokenId
-        tokenPrice.save()
-      }
-    }
-  }
-
-  // Update pool liquidity
-
-  let liquidity = BigDecimal.fromString('0')
-  let denormWeight = BigDecimal.fromString('0')
-
-  for (let i: i32 = 0; i < tokensList.length; i++) {
-    let tokenPriceId = tokensList[i].toHexString()
-    let tokenPrice = TokenPrice.load(tokenPriceId)
-    if (tokenPrice !== null) {
-      let poolTokenId = id.concat('-').concat(tokenPriceId)
-      let poolToken = PoolToken.load(poolTokenId)
-      if (poolToken.denormWeight.gt(denormWeight)) {
-        denormWeight = poolToken.denormWeight
-        liquidity = tokenPrice.price.times(poolToken.balance).div(poolToken.denormWeight).times(pool.totalWeight)
-      }
-    }
-  }
-
-  pool.liquidity = liquidity
-  pool.save()
-}
+import {
+  hexToDecimal,
+  bigIntToDecimal,
+  tokenToDecimal,
+  createPoolShareEntity,
+  createPoolTokenEntity,
+  updatePoolLiquidity,
+  saveTransaction
+} from './helpers'
 
 /************************************
  ********** Pool Controls ***********
@@ -217,20 +31,7 @@ export function handleSetSwapFee(event: LOG_CALL): void {
   pool.swapFee = swapFee
   pool.save()
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'setSwapFee'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'setSwapFee')
 }
 
 export function handleSetController(event: LOG_CALL): void {
@@ -240,20 +41,7 @@ export function handleSetController(event: LOG_CALL): void {
   pool.controller = controller
   pool.save()
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'setController'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'setController')
 }
 
 export function handleSetPublicSwap(event: LOG_CALL): void {
@@ -263,20 +51,7 @@ export function handleSetPublicSwap(event: LOG_CALL): void {
   pool.publicSwap = publicSwap
   pool.save()
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'setPublicSwap'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'setPublicSwap')
 }
 
 export function handleFinalize(event: LOG_CALL): void {
@@ -309,20 +84,7 @@ export function handleFinalize(event: LOG_CALL): void {
   factory.finalizedPoolCount = factory.finalizedPoolCount + 1
   factory.save()
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'finalize'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'finalize')
 }
 
 export function handleRebind(event: LOG_CALL): void {
@@ -363,20 +125,7 @@ export function handleRebind(event: LOG_CALL): void {
 
   updatePoolLiquidity(poolId)
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'rebind'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'rebind')
 }
 
 export function handleUnbind(event: LOG_CALL): void {
@@ -396,20 +145,7 @@ export function handleUnbind(event: LOG_CALL): void {
   pool.save()
   store.remove('PoolToken', poolTokenId)
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'unbind'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'unbind')
 }
 
 export function handleGulp(event: LOG_CALL): void {
@@ -433,20 +169,7 @@ export function handleGulp(event: LOG_CALL): void {
 
   updatePoolLiquidity(poolId)
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'gulp'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'gulp')
 }
 
 /************************************
@@ -469,20 +192,7 @@ export function handleJoinPool(event: LOG_JOIN): void {
 
   updatePoolLiquidity(poolId)
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'join'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'join')
 }
 
 export function handleExitPool(event: LOG_EXIT): void {
@@ -501,20 +211,7 @@ export function handleExitPool(event: LOG_EXIT): void {
 
   updatePoolLiquidity(poolId)
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'exit'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'exit')
 }
 
 /************************************
@@ -570,20 +267,7 @@ export function handleSwap(event: LOG_SWAP): void {
   swap.timestamp = event.block.timestamp.toI32()
   swap.save()
 
-  let tx = event.transaction.hash.toHexString().concat('-').concat(event.logIndex.toString())
-  let transaction = Transaction.load(tx)
-  if (transaction == null) {
-    transaction = new Transaction(tx)
-  }
-  transaction.event = 'swap'
-  transaction.poolAddress = event.address.toHex()
-  transaction.userAddress = event.transaction.from.toHex()
-  transaction.gasUsed = event.transaction.gasUsed.toBigDecimal()
-  transaction.gasPrice = event.transaction.gasPrice.toBigDecimal()
-  transaction.tx = event.transaction.hash
-  transaction.timestamp = event.block.timestamp.toI32()
-  transaction.block = event.block.number.toI32()
-  transaction.save()
+  saveTransaction(event, 'swap')
 }
